+++ CHANGELOG.rst
@@ -1,4 +1,5 @@
 3.1.0
+
 - Add on_grouping_key for custom grouping logic
 - Add Django, Flask and WSGI middleware for Python 3
 - Allow Django settings.py config override